requirements.txt
antlr4-python3-runtime==4.9.3
certifi==2022.9.14
charset-normalizer==2.1.1
colorama
contourpy==1.0.5
cycler
diffusers
filelock
fonttools==4.37.2
fsspec
gmpy2
huggingface_hub
hydra-core==1.3.2
idna
imagecodecs
imageio
importlib-metadata
Jinja2
kiwisolver
lazy_loader
lightning-utilities
lpips==0.1.4
MarkupSafe
matplotlib==3.6.0
mkl-fft
mkl-random
mkl-service==2.4.0
mpmath
networkx
numpy==1.23.3
omegaconf==2.3.0
packaging==21.3
Pillow==9.2.0
ply==3.11
pyparsing
PyQt5==5.15.10
PyQt5-sip
python-dateutil
PyYAML==6.0
regex
requests==2.28.1
safetensors
scikit-image
scipy
sip
six
sympy
tifffile
tomli
torch==2.2.0
torch-fidelity==0.3.0
torchaudio==2.2.0
torchmetrics
torchvision==0.17.0
tornado
tqdm==4.64.1
triton==2.2.0
typing_extensions==4.11.0
unicodedata2
urllib3==1.26.12
zipp
